Fix Overalls error from floating-point points

+++ shared/src/commonMain/kotlin/com/diegopalvarez/oreplay/domain/model/OverallResult.kt
@@ -15,7 +15,7 @@ data class OverallResult(
     val contributory: Boolean?,
     val timeSeconds: Duration,
     val timeBehind: Duration?,
-    val pointsFinal: Long,
-    val pointsBehind: Long?,
+    val pointsFinal: Double,
+    val pointsBehind: Double?,
     val note: String?,
 )
